services/ticket-service/requirements.txt
fastapi==0.128.1
PyJWT==2.8.0
qrcode==8.2
starlette==0.50.0
uvicorn[standard]==0.40.0
Pillow>=10.0.0
